requirements.txt
dash==2.16.1
databricks-sql-connector==3.1.1
python-dotenv==1.0.1
